place all stack-dependent cache folders under one root folder
This way the cache can be reliably cleared without having to maintain a
list of all the different cache directories between multiple files.

This also adds a `elixir_cache_path` and a `erlang_cache_path` functions
for allowing us to keep to track of the Elixir/Erlang caches in one
central place.

fixes #197

https://github.com/HashNuke/heroku-buildpack-elixir/issues/197